Fixes for upstream stdlib

+++ src/constants.rs
@@ -26,7 +26,7 @@ pub static UTXO_SYNC_N_BLOCKS: uint = 500;
 pub static BLOCKCHAIN_N_FULL_BLOCKS: uint = 100;
 
 /// The save-to-disk frequency in s
-pub static SAVE_FREQUENCY: i32 = 600; // 10 minutes
+pub static SAVE_FREQUENCY: i64 = 600; // 10 minutes
 
 /// Default peer address
 pub static DEFAULT_PEER_ADDR: &'static str = "localhost";
